Extract cart components/constants out of the route module
The cart quantity stepper was exported from routes/cart.tsx purely so a
test could import it, which dragged the route's server-only session
helpers (process.env at load) into the browser test bundle and forced a
process.env shim in vitest.browser.config.ts.

Move QuantityPicker into app/components/cart/quantity-picker.tsx and the
shared form intents into app/lib/cart-actions.ts (no server imports), so
the route imports them like everything else. Relocate the test beside the
component and drop the vitest config shim.
